==> 0.10_PIDs.ipynb <==
import os
import pandas as pd
import numpy as np
import csv
from collections import Counter

# 0. Get all unique ids

# get files
folder_in = '/path/to/9_FINAL/data/machine_learning/two_class/distance/merged_with_contexts_'
# os.listdir(folder_in)

provids = []

# get all unique ids
for i in range(1, 39):
    folder = folder_in + str(i) + '/'
    
    # get all files in the folder
    files = os.listdir(folder)
    
    for file in files:
        
        # read data and get provid
        data = pd.read_csv(folder + file, 
                          sep=";",
                          index_col=0)
        
        # debug
        # print(data.provid.tolist())
        
        # extend provids with provid
        provids.extend(data.provid.tolist())
        # print(provids)
        
    print(f'Finished with folder {str(i)}')

# get counter from list 
counter_provids = Counter(provids)

# provids
counter_provids

# export counter to csv
with open('counter_provids_1.csv','w') as csvfile:
    fieldnames=['provid','count']
    writer=csv.writer(csvfile)
    writer.writerow(fieldnames)
    for key, value in counter_provids.items():
        row = str(key) + ';' + str(value)
        writer.writerow(row)

provids = []

# get all unique ids
for i in range(39, 78):
    folder = folder_in + str(i) + '/'
    
    # get all files in the folder
    files = os.listdir(folder)
    
    for file in files:
        
        # read data and get provid
        data = pd.read_csv(folder + file, 
                          sep=";",
                          index_col=0)
        
        # debug
        # print(data.provid.tolist())
        
        # extend provids with provid
        provids.extend(data.provid.tolist())
        # print(provids)
        
    print(f'Finished with folder {str(i)}')

# get counter from list 
counter_provids = Counter(provids)

# export counter to csv
with open('counter_provids_2.csv','w') as csvfile:
    fieldnames=['provid','count']
    writer=csv.writer(csvfile)
    writer.writerow(fieldnames)
    for key, value in counter_provids.items():
        row = str(key) + ';' + str(value)
        writer.writerow(row)

# read counter 1 into memory
counter_1 = pd.read_csv('counter_provids_1.csv',
                       sep=";")
counter_1

# export data
counter_1.to_csv('counter_provids_1.csv',
                 sep=";", index=False)

# read counter 2 into memory
counter_2 = pd.read_csv('counter_provids_2.csv',
                       sep=";")
counter_2

counter_2.to_csv('counter_provids_2.csv',
                 sep=";", index=False)

# get overlap of data
counter_overlap = counter_1['provid'].to_list()
counter_2[counter_2['provid'].isin(counter_1['provid'].to_list())] # none

# get different number of unique provids
len(counter_2) + len(counter_1)

# sort counter_1 by count
counter_1.sort_values(by='count', ascending=False,
                     inplace=True)

# sort counter_2 by count
counter_2.sort_values(by='count', ascending=False, 
                      inplace=True)

# get sum of both counts
(counter_1['count'].sum()) + (counter_2['count'].sum())

# show in which URLs tuples only appear once
len(counter_1[counter_1['count'] == 1]) + len(counter_2[counter_2['count'] == 1])

# top_10 provids of both counters
top_10  = pd.concat([counter_1.iloc[:10,:],
                     counter_2.iloc[:10,:]])
top_10.sort_values(by='count',
                  ascending=False,
                  inplace=True)
top_10 = top_10.iloc[:10,:]

# get array of provid
top_10_provids = top_10.provid.to_list()

context_folder = '/path/to/1_WebisALOD/contextsdb_files/sentencesdb.s'

os.listdir('/path/to/INTENSO/1_WebisALOD/contextsdb_files/')

# test data
pd.read_csv(context_folder + 'sentencesdb.s222.csv',
           sep=",")

# test data
data = pd.read_csv(context_folder + '222.csv',
           sep=",")
data.iloc[0,2]

top_10_provids

# loop through all contexts to get clear text of pld and sentence of top 10
counts=0
for i in range(1,457):
    
    filename = context_folder + str(i) + '.csv'
    print(filename)
    # read data
    data=pd.read_csv(filename,
                     sep=",")
    
    # check whether provid is contained
    data = data[data.provid.isin(top_10_provids)]
    
    if len(data) > 0:
        counts += len(data)
        
        print(f'Found {len(data)} matches')
        
        for j in range(0, len(data)):
            print(f'Provid: {data.iloc[j,1]}')
            print(f'URL: {data.iloc[j,3]}')
            print('Sentence:')
            print(data.iloc[j,2])
    print('--------------')
    if counts == 10:
        break
