==> 0.11_Get_Annotator_Agreement.ipynb <==
import pandas as pd
import numpy as np

from statsmodels.stats.inter_rater import fleiss_kappa

# 0. Get data

# get sample 1 
sample_1 = pd.read_csv("./mturk/webisa_0_sample_500_mturk_output.csv",
                      sep=",")
sample_1.head()

# too many columns to handle, delete all unnecessary columns
sample_1.drop(columns=['HITId', 'HITTypeId', 'Title', 'Description',
                      'Keywords', 'Reward', 'CreationTime', 'MaxAssignments',
                      'RequesterAnnotation', 'AssignmentDurationInSeconds', 
                      'AutoApprovalDelayInSeconds', 'Expiration',
                      'NumberOfSimilarHITs', 'LifetimeInSeconds'], inplace=True)

# too many columns to handle, delete all unnecessary columns
sample_1.drop(columns=['AssignmentId', 'LifetimeApprovalRate',
                      'Last30DaysApprovalRate', 'Last7DaysApprovalRate',
                      'Approve', 'Reject'], inplace=True)

sample_1.shape

# drop all columns with Input.r_ pattern
cols = [c for c in sample_1.columns if c[:8] != 'Input.r_'] #if c.lower()[:6] != 'input.r_']
sample_1 = sample_1[cols]

# drop first few columns
sample_1.drop(columns=['AssignmentStatus', 'AcceptTime',
                      'SubmitTime', 'AutoApprovalTime', 'ApprovalTime',
                      'RejectionTime', 'RequesterFeedback', 'WorkTimeInSeconds'],
             inplace=True)

sample_1.head()

data = []
# get the right answer per workerid and input
for index, row in sample_1.iterrows():
    
    # loop through all answerst per workerId
    for i in range(0, 20):
        
        # get 
        column_name = 'Input.id_' + str(i)
        
        # get id of question
        question_id = str(row[column_name])
        
        # get answer
        answer = 'Answer.' + question_id
        answer = row[answer]
        
        data.append([row['WorkerId'], question_id, answer])
        
        

# all data
all_data = pd.DataFrame(data,  columns=['Annotator', 'id', 'label'])
all_data

def get_data(str_folder):
    sample = pd.read_csv(str_folder,
                      sep=",")
    
    # too many columns to handle, delete all unnecessary columns
    sample.drop(columns=['HITId', 'HITTypeId', 'Title', 'Description',
                      'Keywords', 'Reward', 'CreationTime', 'MaxAssignments',
                      'RequesterAnnotation', 'AssignmentDurationInSeconds', 
                      'AutoApprovalDelayInSeconds', 'Expiration',
                      'NumberOfSimilarHITs', 'LifetimeInSeconds',
                      'AssignmentId', 'LifetimeApprovalRate',
                      'Last30DaysApprovalRate', 'Last7DaysApprovalRate',
                      'Approve', 'Reject', 
                      'AssignmentStatus', 'AcceptTime',
                      'SubmitTime', 'AutoApprovalTime', 'ApprovalTime',
                      'RejectionTime', 'RequesterFeedback', 'WorkTimeInSeconds'], inplace=True)
    
    # drop all columns with Input.r_ pattern
    cols = [c for c in sample.columns if c[:8] != 'Input.r_'] #if c.lower()[:6] != 'input.r_']
    sample = sample[cols]
    
    data = []
    # get the right answer per workerid and input
    for index, row in sample.iterrows():
    
        # loop through all answerst per workerId
        for i in range(0, 20):
        
            # get 
            column_name = 'Input.id_' + str(i)
        
            # get id of question
            question_id = str(row[column_name])
        
            # get answer
            answer = 'Answer.' + question_id
            answer = row[answer]
        
            data.append([row['WorkerId'], question_id, answer])
        
    return data

# loop through all folders
all_data = []

folder = "./mturk/webisa_" 
#10,20
for j in [0,1,2,3,5,10,20]:
    
    folder_name = folder + str(j) + '_sample_500_mturk_output.csv'
    print(folder_name)
    
    data = get_data(folder_name)
    
    all_data.extend(data)

# convert to dataframe for easier indexing
all_data = pd.DataFrame(all_data, columns=['Annotator', 'id', 'label'])
all_data

# show whether function worked
all_data['Key'] = all_data['Annotator'] + ' ' + all_data['id'] + ' ' + all_data['label']
all_data.Key.value_counts()

duplicates = pd.DataFrame(all_data.Key.value_counts())
duplicates = duplicates[duplicates['Key'] > 1].reset_index()
duplicates.columns = ['Key', 'Counts']
duplicates

# get distinct annotators
all_data.Annotator.value_counts()

# remove duplicate keys from dataframe
all_data = all_data[~all_data['Key'].isin(duplicates.Key.to_list())]
len(all_data)

# get all data id
all_data.id.value_counts()

# delete Keys as column
del all_data['Key']

# show how often the data was filled
all_data.Annotator.value_counts()

# pivot the values
all_data_pivot = all_data.pivot_table(values='label',
                                     index='id',
                                     columns='Annotator',
                                     aggfunc='first')

all_data_pivot

def pad_with_num_values(row)

all_data_pivot
