==> 6.4_Check_pid_negative_examples.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

# read in negative examples
negative_examples = pd.read_csv('/path/to/9_FINAL/data/all_negative_examples.csv', sep=";")
del negative_examples['Unnamed: 0']
negative_examples.head()

# 1. pidspread

# show how many pids are the "norm"
negative_examples['pidspread'].describe() 
# subclass mean: 11.62
# transitive subclass: 10.98
# type: 4.73
# transitive type: 3.29
# negative examples: 2.59 -> way less than normal

# show which pids are contained how often
negative_examples['pidspread'].value_counts()

# 2. pid distribution

# replace unnecessary signs
negative_examples['modifications'] = negative_examples['modifications'].replace('\]|,\s|\'|\[', '', regex=True)
negative_examples

# get all different pids
negative_examples['pids_sanitized'] = negative_examples['modifications'].str.split('p')

all_pids = []
for index, row in negative_examples.iterrows():
    for element in row['pids_sanitized']:
        if element not in all_pids:
            all_pids.append(element)

all_pids.remove('') # remove empty element

len(all_pids) # in total 58 different pids in subclasses

# append number of occurrences as new column to negative classes dataframe
def pid_exists(row, pid):
    if pid in row['modifications']:
        return 1
    else:
        return 0

occurences = [] # no of occurences of pids

for pid in all_pids:
    
    negative_examples[pid] = negative_examples.apply(lambda x: pid_exists(x, pid), axis=1) # create new column with one hot encoded occurence
    print('pid: {} || total occurence: {}'.format(pid, negative_examples[pid].sum())) # for debugging
    occurences.append(negative_examples[pid].sum())
