==> 8.1_ML_One_Hot_CV_Downsampled.ipynb <==
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.metrics import confusion_matrix, plot_confusion_matrix, f1_score
from sklearn.model_selection import cross_val_score, KFold # for cross validation
from sklearn import preprocessing
from sklearn.neural_network import MLPClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn import metrics # for f1 macro in cross validation

import xgboost as xgb

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import itertools
from helper import util_visualizations, util_ml

# 0. Load Data

## 0.1 Training

# read data
train = pd.read_csv('/path/to/9_FINAL/data/machine_learning/one_hot_encoded/train/downsampled_training_data.csv', sep=";")
del train['Unnamed: 0']

# get label
y_train = train['label']

# get variables
X_train = train.copy()
del X_train['label']
X_train.shape

X_train

# only take one hot encoded columns
X_train_one_hot = X_train.drop(['instance', 'class', 'frequency', 'pidspread', 'pldspread', 'id', 'pids'], axis=1)
X_train_one_hot.head()

# only take one hot encoded columns
X_train_one_hot_more = X_train.drop(['instance', 'class', 'id', 'pids'], axis=1) 
X_train_one_hot_more.head()

# scale
min_max_scaler = preprocessing.MinMaxScaler()
X_train_one_hot_more_scaled = pd.DataFrame(min_max_scaler.fit_transform(X_train_one_hot_more))
X_train_one_hot_more_scaled.head()

# only take one hot encoded columns
X_train_one_hot_frequency = X_train[['frequency', 'pidspread', 'pldspread']]# TODO: frequency
X_train_one_hot_frequency.head()

# scale
min_max_scaler = preprocessing.MinMaxScaler()
X_train_one_hot_frequency_scaled = pd.DataFrame(min_max_scaler.fit_transform(X_train_one_hot_frequency))
X_train_one_hot_frequency_scaled.head()

## 0.2 Testing

# read test data
test = pd.read_csv('/path/to/9_FINAL/data/machine_learning/one_hot_encoded/test/test_data.csv', sep=";")
del test['Unnamed: 0']

# get label
y_test = test['label']

# get variables
X_test = test.copy()
del X_test['label']
X_test.shape

# only take one hot encoded columns
X_test_one_hot = X_test.drop(['instance', 'class', 'frequency', 'pidspread', 'pldspread', 'id', 'pids'], axis=1)
X_test_one_hot.head()

# only take one hot encoded columns
X_test_one_hot_more = X_test.drop(['instance', 'class',  'id', 'pids'], axis=1)
X_test_one_hot_more.head()

# scale
min_max_scaler = preprocessing.MinMaxScaler()
X_test_one_hot_more_scaled = pd.DataFrame(min_max_scaler.fit_transform(X_test_one_hot_more))
X_test_one_hot_more_scaled.head()

# only take one hot encoded columns
X_test_one_hot_frequency = X_test[['frequency', 'pidspread', 'pldspread']]
X_test_one_hot_frequency.head()

# scale
min_max_scaler = preprocessing.MinMaxScaler()
X_test_one_hot_frequency_scaled = pd.DataFrame(min_max_scaler.fit_transform(X_test_one_hot_frequency))
X_test_one_hot_frequency_scaled.head()

# 1. Naive Bayes

## 1.1 Only one hot encoded columns

# fit NB model
nb = MultinomialNB()
util_ml.get_cross_validated_confusion_matrix(nb, X_train_one_hot, y_train)

## 1.2 Only one hot encoded columns with frequencies

# fit NB model
nb = MultinomialNB()
util_ml.get_cross_validated_confusion_matrix(nb, X_train_one_hot_more, y_train)

## 1.3 Only with frequencies

# fit NB model
nb = MultinomialNB()
util_ml.get_cross_validated_confusion_matrix(nb, X_train_one_hot_frequency, y_train)

## 1.4 Scaled one hot encoded pids and frequencies

# fit NB model
nb = MultinomialNB()
util_ml.get_cross_validated_confusion_matrix(nb, X_train_one_hot_more_scaled, y_train)

## 1.5 Frequencies normalized

# fit NB model
nb = MultinomialNB()
util_ml.get_cross_validated_confusion_matrix(nb, X_train_one_hot_frequency_scaled, y_train)

# 2. Decision Trees

## 2.1 Only one hot encoded columns

# fit DT model
dt = DecisionTreeClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(dt, X_train_one_hot, y_train)

## 2.2 Only one hot encoded columns with frequencies

# fit DT model
dt = DecisionTreeClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(dt, X_train_one_hot_more, y_train)

## 2.3 Only with frequencies

# fit DT model
dt = DecisionTreeClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(dt, X_train_one_hot_frequency, y_train)

## 2.4 Scaled one hot encoded pids and frequencies

# fit DT model
dt = DecisionTreeClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(dt, X_train_one_hot_more_scaled, y_train)

## 2.5 Frequencies normalized

# fit DT model
dt = DecisionTreeClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(dt, X_train_one_hot_frequency_scaled, y_train)

# 3. Random Forest

## 3.1 Only one hot encoded columns

# fit random forest model
rf = RandomForestClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(rf, X_train_one_hot, y_train)

## 3.2 Only one hot encoded columns with frequencies

# fit random forest model
rf = RandomForestClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(rf, X_train_one_hot_more, y_train)

## 3.3 Only with frequencies

# fit random forest model
rf = RandomForestClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(rf, X_train_one_hot_frequency, y_train)

## 3.4 Scaled one hot encoded pids and frequencies

# fit random forest model
rf = RandomForestClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(rf, X_train_one_hot_more_scaled, y_train)

## 3.5 Frequencies normalized

# fit random forest model
rf = RandomForestClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(rf, X_train_one_hot_frequency_scaled, y_train)

# 4. Neural Network

## 4.1 Only one hot encoded columns

# fit nnet
nnet = MLPClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(nnet, X_train_one_hot, y_train)

## 4.2 Only one hot encoded columns with frequencies

# fit nnet
nnet = MLPClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(nnet, X_train_one_hot_more, y_train)

## 4.3 Only with frequencies

# fit nnet
nnet = MLPClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(nnet, X_train_one_hot_frequency, y_train)

## 4.4 Scaled one hot encoded pids and frequencies

# fit nnet
nnet = MLPClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(nnet, X_train_one_hot_more_scaled, y_train)

## 4.5 Frequencies normalized

# fit nnet
nnet = MLPClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(nnet, X_train_one_hot_frequency_scaled, y_train)

# 5. Logistic Regression

## 5.1 Only one hot encoded columns

# fit LogReg
log_reg = LogisticRegression(random_state=88)
util_ml.get_cross_validated_confusion_matrix(log_reg, X_train_one_hot, y_train)

## 5.2 Only one hot encoded columns with frequencies

# fit LogReg
log_reg = LogisticRegression(random_state=88)
util_ml.get_cross_validated_confusion_matrix(log_reg, X_train_one_hot_more, y_train)

## 5.3 Only with frequencies

# fit LogReg
log_reg = LogisticRegression(random_state=88)
util_ml.get_cross_validated_confusion_matrix(log_reg, X_train_one_hot_frequency, y_train)

## 5.4 Scaled one hot encoded pids and frequencies

# fit LogReg
log_reg = LogisticRegression(random_state=88)
util_ml.get_cross_validated_confusion_matrix(log_reg, X_train_one_hot_more_scaled, y_train)

## 5.5 Frequencies normalized

# fit LogReg
log_reg = LogisticRegression(random_state=88)
util_ml.get_cross_validated_confusion_matrix(log_reg, X_train_one_hot_frequency_scaled, y_train)

# 6. XGBoost

## 6.1 Only one hot encoded columns

# fit xgboost
xg = xgb.XGBClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(xg, X_train_one_hot, y_train)

## 6.2 Only one hot encoded columns with frequencies

# fit xgboost
xg = xgb.XGBClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(xg, X_train_one_hot_more, y_train)

## 6.3 Only with frequencies

# fit xgboost
xg = xgb.XGBClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(xg, X_train_one_hot_frequency, y_train)

## 5.4 Scaled one hot encoded pids and frequencies

# fit xgboost
xg = xgb.XGBClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(xg, X_train_one_hot_more_scaled, y_train)

## 6.5 Frequencies normalized

# fit xgboost
xg = xgb.XGBClassifier(random_state=88)
util_ml.get_cross_validated_confusion_matrix(xg, X_train_one_hot_frequency_scaled, y_train)
