==> 9.5_Get_count_columns.ipynb <==
import pandas as pd
from helper import util_pids

# get all pids as list
train = pd.read_csv('/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/training/training_0.csv',sep=";")
train = train.drop(['Unnamed: 0', 'Unnamed: 0.1', 'instance', 'class', 'frequency', 'pidspread', 'pldspread', 'id', 'pids', 'label'], axis=1)

# different pids
pids = train.columns.to_list()
pids[:5]

# sanity check
pids[-5:]

# calculate the counts of the data and save the files
for x in range(5):
    # load data
    train = pd.read_csv('/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/training/training_'+str(x)+'.csv',sep=";")
    
    # debug
    del train['Unnamed: 0']
    del train['Unnamed: 0.1']
    
    # reserve labels
    labels = train['label']
    
    # only keep instance, class, frequency, pidspread, pldspread & pids from train
    train = train[['instance', 'class', 'frequency', 'pidspread', 'pldspread', 'id', 'pids']]
    
    # calculate count based features
    count_train = util_pids.append_count_pids_to_data(train, pids, 'pids')
    
    # append label to train
    count_train['label'] = labels
    print(count_train.head())
    
    # export data for future uses
    count_train.to_csv('/path/to/9_FINAL/data/machine_learning/count_based/cross_validation/training/training_'+str(x)+'.csv', sep=";")

# sanity check
count_train.shape

# calculate the counts of the data and save the files for validation splits
for x in range(5):
    # load data
    validation = pd.read_csv('/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/validation/val_fold_'+str(x)+'.csv',sep=";")

    # debug
    del validation['Unnamed: 0']
    
    # reserve labels
    labels = validation['label']
    
    # only keep instance, class, frequency, pidspread, pldspread & pids from train
    validation = validation[['instance', 'class', 'frequency', 'pidspread', 'pldspread', 'id', 'pids']]
    
    # calculate count based features
    count_validation = util_pids.append_count_pids_to_data(validation, pids, 'pids')
    
    # append label to train
    count_validation['label'] = labels
    print(count_validation.shape)
    print(count_validation.head())
    
    # export data for future uses
    count_validation.to_csv('/path/to/9_FINAL/data/machine_learning/count_based/cross_validation/validation/validation_'+str(x)+'.csv', sep=";")

count_validation.shape

# get goldstandard
goldstandard = pd.read_csv('/path/to/9_FINAL/data/goldstandard/goldstandard_count.csv',sep=";")
goldstandard.columns

count_validation = pd.read_csv('/path/to/9_FINAL/data/machine_learning/count_based/cross_validation/validation/validation_0.csv', sep=";")
count_validation.columns
