==> 9.7_Normally_distributed_folds.ipynb <==
import pandas as pd
import os

# 1. Stack datafolds with original class distribution for one hot encoded variables

# get files
negative = '/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/negative/'
positive = '/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/positive/'

negative_files = os.listdir(negative)
positive_files = os.listdir(positive)

# get length of positives
file = 'fold_positives_0.csv'
data = pd.read_csv(positive + file, sep=";")
print(len(data)*4)

# get length of negatives in training 
file = 'fold_negatives_0.csv'
data = pd.read_csv(negative + file, sep=";")
print(len(data)*4)

# complete theoretic length of data 
len(data)*4 + 91384 # -> too big!

# loop through all folds
for x in range(5):
    
    all_positives=pd.DataFrame()
    do_not_take_positive = str(x).join(['fold_positives_', '.csv'])
    
    # loop through all positives
    for file in positive_files:
        if file != do_not_take_positive:
            data = pd.read_csv(positive + file, sep=";")
            del data['Unnamed: 0']
            
            # only get 50% since size is too big
            sample = int(0.5 * len(data))
            data = data.sample(sample)
            
            # concat data to data from before
            all_positives = pd.concat([all_positives, data])
            del data
    
    all_negatives = pd.DataFrame()
    do_not_take_negative = str(x).join(['fold_negatives_', '.csv'])
    
    # loop through all negatives
    for negative_file in negative_files:
        if negative_file != do_not_take_negative:
            data = pd.read_csv(negative + negative_file, sep=";")
            del data['Unnamed: 0']
            
            # only get 50% since size is too big
            sample = int(0.5 * len(data))
            data = data.sample(sample)
            
            # concat data to before
            all_negatives = pd.concat([all_negatives, data])
            del data
            
    # stack data
    training_fold = pd.concat([all_positives, all_negatives])
    training_fold = training_fold.sample(frac=1).reset_index(drop=True)
    
    # sanity check
    print('Training {}'.format(len(training_fold)))
    
    # export
    file_out = 'training_' + str(x) + '.csv'
    training_fold.to_csv('/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/training_normal_distribution/'+file_out,sep=";")

# loop through all folds
for x in range(5):
    
    all_positives=pd.DataFrame()
    do_not_take_positive = str(x).join(['fold_positives_', '.csv'])
    
    # loop through all positives
    for file in positive_files:
        if file != do_not_take_positive:
            data = pd.read_csv(positive + file, sep=";")
            del data['Unnamed: 0']
            
            # only get 20% since size is too big
            sample = int(0.2 * len(data))
            data = data.sample(sample)
            
            # concat data to data from before
            all_positives = pd.concat([all_positives, data])
            del data
    
    all_negatives = pd.DataFrame()
    do_not_take_negative = str(x).join(['fold_negatives_', '.csv'])
    
    # loop through all negatives
    for negative_file in negative_files:
        if negative_file != do_not_take_negative:
            data = pd.read_csv(negative + negative_file, sep=";")
            del data['Unnamed: 0']
            
            # only get 20% since size is too big
            sample = int(0.2 * len(data))
            data = data.sample(sample)
            
            # concat data to before
            all_negatives = pd.concat([all_negatives, data])
            del data
            
    # stack data
    training_fold = pd.concat([all_positives, all_negatives])
    training_fold = training_fold.sample(frac=1).reset_index(drop=True)
    
    # sanity check
    print('Training {}'.format(len(training_fold)))
    
    # export
    file_out = 'training_' + str(x) + '.csv'
    training_fold.to_csv('/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/training_fraction_normal_distribution/'+file_out,sep=";")
