==> 9.12_Get_fraction_count_based.ipynb <==
import pandas as pd

out_folder = '/path/to/9_FINAL/data/machine_learning/count_based/cross_validation/training_fraction_normal_distribution/training_'

validation_folder = '/path/to/9_FINAL/data/machine_learning/count_based/cross_validation/validation/validation_'

# get ids in one hot encoded to get the same ids
training_file = '/path/to/9_FINAL/data/machine_learning/two_class/one-hot-ecoding/train/cross_validation/training_fraction_normal_distribution/training_'
for i in range(0,5):
    # read training file
    training = training_file + str(i) + '.csv'
    
    # get id of training data
    training_data = pd.read_csv(training,  sep=";")
    training_data = training_data.id.to_list()
    
    # debug
    print(f'Length file one hot encoded {len(training_data)}')
    
    # get the training fraction
    training_fraction = pd.DataFrame()
    # loop through all validation files to get the right ids
    
    # get indices of validation sets to check
    sets_to_check = [0,1,2,3,4]
    sets_to_check.remove(i) # since validation set cannot be included, excluded these observations
    
    for j in sets_to_check:
        validation_file_name = validation_folder + str(j) + '.csv'
        print(validation_file_name)
        validation_file = pd.read_csv(validation_file_name, sep=";", index_col=0)
        
        # only get the ids of the fraction
        validation_file = validation_file[validation_file.id.isin(training_data)]
        
        if len(validation_file) > 0 :
            training_fraction = pd.concat([training_fraction, validation_file], ignore_index=True)
    
    # debug
    print(f'Length of found matches {len(training_fraction)}')
    
    # export file
    training_fraction.to_csv(out_folder+str(i)+'.csv', sep=";")
