==> 12.3_Count_Three_class_subclass.ipynb <==
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.tree import DecisionTreeClassifier

import xgboost as xgb

from helper import util_three_class

# 0. Get folder names for train and validation

folder_train = '/path/to/9_FINAL/data/machine_learning/count_based/three_class/train/downsampled_to_subclass/train_'

folder_validation = '/path/to/9_FINAL/data/machine_learning/one_hot_encoded/three_class/validation/validation_'

# 1. Logistic Regression

## 1.1 Count columns

lr = LogisticRegression(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                lr)

## 1.2 Scaled count columns

lr = LogisticRegression(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                lr, scale=True)

## 1.3 Count columns + frequencies

lr = LogisticRegression(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                lr)

## 1.4 Scaled count columns + frequencies

lr = LogisticRegression(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                lr, scale=True)

# 2. Naive Bayes

## 2.1 Count encoded columns

nb = MultinomialNB()

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                nb)

## 2.2 Scaled Count encoded columns

nb = MultinomialNB()

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                nb, scale=True)

## 2.3 Count columns + frequencies

nb = MultinomialNB()

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                nb)

## 2.4 Scaled count columns + frequencies

nb = MultinomialNB()

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                nb,scale=True)

# 3. XGBoost

## 3.1 Count columns

xg = xgb.XGBRFClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                xg)

## 3.2 Scaled Count columns

xg = xgb.XGBRFClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                xg, scale=True)

## 3.3 Count columns + frequencies

xg = xgb.XGBRFClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                xg)

## 3.4 Scaled Count columns + frequencies

xg = xgb.XGBRFClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                xg, scale=True)

# 4. Random Forest

## 4.1 Count encoded columns

rf = RandomForestClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                rf)

## 4.2 Scaled Count encoded columns

rf = RandomForestClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                rf, scale=True)

## 4.3 Count columns + frequencies

rf = RandomForestClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class','id', 'pids'],
                rf)

## 4.4 scaled Count columns + frequencies

rf = RandomForestClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class','id', 'pids'],
                rf, scale=True)

# 5. MLPClassifier

## 5.1 Count encoded columns

nnet = MLPClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                nnet)

## 5.2 Scaled Count encoded columns

nnet = MLPClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                nnet, scale=True)

## 5.3 Count encoded columns + frequencies

nnet = MLPClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                nnet)

## 5.4 Scaled Count encoded columns + frequencies

nnet = MLPClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                nnet, scale=True)

# 6. DecisionTreeClassifier

## 6.1 Count encoded columns

dt = DecisionTreeClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                dt)

## 6.2 Scaled Count encoded columns

dt = DecisionTreeClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'frequency', 'pidspread', 
                 'pldspread', 'id', 'pids'],
                dt, scale=True)

## 6.3 Only count encoded columns + frequencies

dt = DecisionTreeClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                dt)

## 6.4 Scaled count encoded columns + frequencies

dt = DecisionTreeClassifier(random_state=42)

util_three_class.cross_validation(folder_validation, folder_train,
                ['instance', 'class', 'id', 'pids'],
                dt, scale=True)
